@@ -17,11 +17,12 @@
   },
 
   "require": {
-    "php": "^7.2",
+    "php": "^7.4|^8.0",
     "ext-json": "*"
   },
 
   "require-dev": {
-    "phpunit/phpunit": ">=8.5.8"
+    "phpunit/phpunit": "^9.5",
+    "phpstan/phpstan": "^0.12.63"
   }
 }
